- minor updates for win32 compat. - define constants as Float to avoid loss of precision warning - Prefix to take masklen as a short.

+++ common/tcanetpp/include/Prefix.hpp
@@ -55,7 +55,7 @@ class Prefix {
         _pfx.addrlen  = 0;
     }
 
-    Prefix ( ipv4addr_t addr, uint8_t len )
+    Prefix ( ipv4addr_t addr, uint16_t len )
     {
         _pfx.addr     = addr;
         _pfx.addrlen  = len;
